refactor: use upstream UnifiedLog analyzeAndValidateRecords [INK-127]
Now that analyzeAndValidateRecords is available on UnifiedLog.java make
use of it by tweaking the access to the method.
